+++ django_mysql/apps.py
@@ -1,5 +1,6 @@
 # -*- coding:utf-8 -*-
 from django.apps import AppConfig
+from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 
 
@@ -13,10 +14,14 @@ class MySQLConfig(AppConfig):
 
     def perform_monkey_patches(self):
         from django.db.backends.mysql.base import DatabaseWrapper
-        from django_mysql.monkey_patches import is_mariadb
+        from django_mysql import monkey_patches
 
         # Fine to patch straight on since it's a cached_property descriptor
-        DatabaseWrapper.is_mariadb = is_mariadb
+        DatabaseWrapper.is_mariadb = monkey_patches.is_mariadb
+
+        # Depends on setting
+        if getattr(settings, 'DJANGO_MYSQL_REWRITE_QUERIES', False):
+            monkey_patches.patch_CursorWrapper_execute()
 
     def add_lookups(self):
         from django.db.models import CharField, TextField