+++ HISTORY.rst
@@ -7,6 +7,9 @@ History
 ------------------
 
 * Pending release
+* Added query-rewriting layer added which allows the use of MySQL query hints
+  such as ``STRAIGHT_JOIN`` via QuerySet methods, as well as adding label
+  comments to track where queries are generated.
 
 0.2.0 (2015-05-14)
 ------------------